requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-socketio[asgi]==5.11.3
PyMySQL==1.1.1
pydantic==2.8.2
pydantic-settings==2.3.4
python-dotenv==1.0.1
